Update faucet UI
