+++ centrifuge-app/src/components/Faucet.tsx
@@ -1,18 +1,53 @@
-import { useWallet } from '@centrifuge/centrifuge-react'
+import { CurrencyBalance, findCurrency } from '@centrifuge/centrifuge-js'
+import { useAddress, useBalances, useWallet } from '@centrifuge/centrifuge-react'
 import { Button, Card, Shelf, Stack, Text } from '@centrifuge/fabric'
+import BN from 'bn.js'
 import * as React from 'react'
+import { useParams } from 'react-router'
+import { useCurrencies } from '../utils/useCurrencies'
+import { usePool } from '../utils/usePools'
 import { FaucetConfirmationDialog } from './Dialogs/FaucetConfirmationDialog'
 
-export const Faucet: React.VFC = () => {
+const MIN_DEVEL_BALANCE = 10
+const MIN_AUSD_BALANCE = 100
+
+export const Faucet = () => {
   const { selectedAccount } = useWallet().substrate
   const [hash, setHash] = React.useState('')
   const [error, setError] = React.useState('')
   const [isLoading, setIsLoading] = React.useState(false)
+  const currencies = useCurrencies()
+  const { pid: poolId } = useParams<{ pid: string }>()
+  const pool = usePool(poolId)
+  const balances = useBalances(useAddress('substrate'))
+
+  const { connectedType } = useWallet()
+  const isTinlakePool = poolId?.startsWith('0x')
+
+  const hasLowNativeBalance =
+    balances && new CurrencyBalance(balances.native.balance, 18).toDecimal().lte(MIN_DEVEL_BALANCE)
+  const poolCurrency = findCurrency(currencies ?? [], pool?.currency?.key)
+  const poolCurrencyBalance =
+    balances?.currencies.find((curr) => curr.currency.symbol === poolCurrency?.symbol)?.balance ?? new BN(0)
+  const hasLowAusdBalance =
+    (poolCurrency &&
+      new CurrencyBalance(poolCurrencyBalance, poolCurrency.decimals).toDecimal().lte(MIN_AUSD_BALANCE)) ||
+    !poolCurrency
+
+  const shouldRenderFaucet =
+    poolCurrency &&
+    connectedType === 'substrate' &&
+    !isTinlakePool &&
+    import.meta.env.REACT_APP_FAUCET_URL &&
+    hasLowNativeBalance &&
+    hasLowAusdBalance
 
   const handleClaim = async () => {
     setIsLoading(true)
     try {
-      const response = await fetch(`${import.meta.env.REACT_APP_FAUCET_URL}?address=${selectedAccount?.address}`)
+      const response = await fetch(
+        `${import.meta.env.REACT_APP_FAUCET_URL}?address=${selectedAccount?.address}&poolId=${pool?.id}`
+      )
       if (response.status !== 200) {
         throw response.text()
       }
@@ -25,7 +60,7 @@ export const Faucet: React.VFC = () => {
     }
   }
 
-  return (
+  return shouldRenderFaucet ? (
     <>
       <FaucetConfirmationDialog
         error={error}
@@ -42,7 +77,7 @@ export const Faucet: React.VFC = () => {
             Faucet
           </Text>
           <Text as="p" variant="heading3">
-            1k {import.meta.env.REACT_APP_COLLATOR_WSS_URL.includes('demo') ? 'DEMO' : 'DEVEL'} and 10k aUSD
+            1k Native and 10k {pool.currency.symbol}
           </Text>
         </Stack>
         <Button loading={isLoading} disabled={isLoading} onClick={handleClaim} variant="primary">
@@ -50,5 +85,5 @@ export const Faucet: React.VFC = () => {
         </Button>
       </Shelf>
     </>
-  )
+  ) : null
 }
