+++ centrifuge-app/src/components/PageWithSideBar.tsx
@@ -1,13 +1,9 @@
-import { CurrencyBalance } from '@centrifuge/centrifuge-js'
-import { useBalances, useWallet, WalletMenu } from '@centrifuge/centrifuge-react'
+import { WalletMenu } from '@centrifuge/centrifuge-react'
 import { Box, Grid, Shelf, Stack } from '@centrifuge/fabric'
 import * as React from 'react'
-import { useParams } from 'react-router-dom'
 import { useTheme } from 'styled-components'
 import { config } from '../config'
-import { useAddress } from '../utils/useAddress'
 import { useIsAboveBreakpoint } from '../utils/useIsAboveBreakpoint'
-import { Faucet } from './Faucet'
 import { Footer } from './Footer'
 import { LoadBoundary } from './LoadBoundary'
 import { LogoLink } from './LogoLink'
@@ -19,9 +15,6 @@ type Props = {
   children?: React.ReactNode
 }
 
-const MIN_DEVEL_BALANCE = 10
-const MIN_AUSD_BALANCE = 100
-
 const TOOLBAR_HEIGHT = 75
 const HEADER_HEIGHT = 56
 const MENU_WIDTH = 80
@@ -30,24 +23,7 @@ export const PAGE_GUTTER = ['gutterMobile', 'gutterTablet', 'gutterDesktop']
 
 export const PageWithSideBar: React.FC<Props> = ({ children, sidebar = true }) => {
   const isMedium = useIsAboveBreakpoint('M')
-  const { connectedType } = useWallet()
-  const { pid: poolId } = useParams<{ pid: string }>()
-  const isTinlakePool = poolId?.startsWith('0x')
-
   const theme = useTheme()
-  const balances = useBalances(useAddress('substrate'))
-  const hasLowDevelBalance =
-    balances && new CurrencyBalance(balances.native.balance, 18).toDecimal().lte(MIN_DEVEL_BALANCE)
-  const aUSD = balances && balances.currencies.find((curr) => curr.currency.key === 'AUSD')
-  const hasLowAusdBalance =
-    (aUSD && new CurrencyBalance(aUSD.balance, aUSD.currency.decimals).toDecimal().lte(MIN_AUSD_BALANCE)) || !aUSD
-
-  const shouldRenderFaucet =
-    connectedType === 'substrate' &&
-    !isTinlakePool &&
-    import.meta.env.REACT_APP_FAUCET_URL &&
-    hasLowDevelBalance &&
-    hasLowAusdBalance
 
   return (
     <Grid
@@ -145,8 +121,6 @@ export const PageWithSideBar: React.FC<Props> = ({ children, sidebar = true }) =
               <Stack mb={9} px={8} gap={4}>
                 <WalletMenu menuItems={[<OnboardingStatus />]} />
               </Stack>
-
-              {shouldRenderFaucet && <Faucet />}
 
               <LoadBoundary>{sidebar}</LoadBoundary>
             </Stack>
